Split up core views
